Add a link to the no. of contributions that goes to the contributions search
The link adds the contributions page id to the URL, so the search will only
show contributions donated to the contribution page of the PCP. Ideally, it
would add the PCP id to the URL, but I couldn't find a way to do it.

If there's 0 contributions or the contributions are actually event registrations
(which doesn't have contributions page), there will be no link.